+++ validator/src/commands/exit/mod.rs
@@ -53,6 +53,13 @@ impl FromClapArgMatches for ExitArgs {
                  default behavior"
             );
         }
+        // Deprecated in v3.1.0
+        if matches.is_present("monitor") {
+            eprintln!(
+                "WARN: The --monitor flag has been deprecated, use \"agave-validator monitor\" \
+                 instead"
+            );
+        }
 
         Ok(ExitArgs {
             force: matches.is_present("force"),
@@ -84,6 +91,7 @@ pub fn command<'a>() -> App<'a, 'a> {
                 .long("monitor")
                 .takes_value(false)
                 .requires("no_wait_for_exit")
+                .hidden(hidden_unless_forced())
                 .help("Monitor the validator after sending the exit request"),
         )
         .arg(